Automated rollback of commit 52212faf50d2d4a6f1aa9c84203c967217df6c22.
*** Reason for rollback ***

Causes NPE in the nightly

*** Original change description ***

C++: Uses new API in transitive linking

We convert between the old and the new API. It's fine to flatten the nested sets here with the conversion since they are flattened anyway. Once everything is converted to the new API, the conversion step would be unnecessary.

RELNOTES:none
PiperOrigin-RevId: 226025709