Linkify all flags in the Command Line Reference
This updates the option html emitter to add anchor links and id tags to
each generated flag in the command line reference. It also fixes its
color to be consistently red in the code block, instead of the standard
Bazel green for links.

Live demo:

https://bazel-docs-staging.netlify.com/versions/master/command-line-reference.html#flag--all_incompatible_changes
https://bazel-docs-staging.netlify.com/versions/master/command-line-reference.html#flag--experimental_oom_more_eagerly_threshold

When not hovering:

![ss](https://user-images.githubusercontent.com/347918/50126797-6b1f0880-023c-11e9-85fa-5d47c7f62a25.png)

When hovering:

![ss](https://user-images.githubusercontent.com/347918/50126805-740fda00-023c-11e9-93c2-d69e719dcf34.png)

and that generates the link:

`https://docs.bazel.build/versions/<version>/command-line-reference.html#flag--experimental_oom_more_eagerly_threshold`

Expanded flags are also linkified:

![ss](https://user-images.githubusercontent.com/347918/50127567-b2f35f00-023f-11e9-8bf4-e39d3937929c.png)

There may be a concern that we don't guarantee that flags are only generated once, and can be duplicated, but because flag definitions don't change, linking to the first instance of the flag is safe.

Fixes https://github.com/bazelbuild/bazel/issues/3502

Change-Id: I6775feaa48558029b943b3206891e8b6860bb5d9
RELNOTES: None.

Closes #6954.

Change-Id: I37057365b0a117e269bdaa56bdec8153fe0a5b85
PiperOrigin-RevId: 226065220