Preserve order of newlyRequestedDeps in SkyFunctionEnvironment.
Construction of newlyRequestedDeps is pushed down to getValuesFromErrorOrDepsOrGraph so that keys can be added during iteration over the Iterable<SkyKey> of requested deps, not after they have already been placed into a HashMap (which changes their order).

An alternate possible implementation would be to switch from HashMap to LinkedHashMap, but this would tie up more memory since the map is retained in the return value (Maps.transformValues is lazy).

RELNOTES: None.
PiperOrigin-RevId: 226345924