Use artifact id and sha1 for cache key type for maven_jar artifacts
Previously the RepositoryCache was unable to provide any
identification of cached entries except for the sha1 value itself,
which makes it hard to verify that a retrieved value is what the user
expects. At least in the case of maven_jar rules, the artifactId
attribute has more user significance than the sha1.

This CL adds a new KeyType in the RepositoryCache called
ID_PREFIXED_SHA1 which takes a key that has a concatenated form of the
maven coordinates as a prefix to the sha1 portion of the key. This new
key type is used for resolving cached maven artifacts (and storing new
ones), which will ensure that the artifact id and sha1 match and are
valid.

https://github.com/bazelbuild/bazel/issues/5144