Change JobsConverter to ResourceConverter type to facilitate auto-tuning in the future.
Behavior of --jobs is the same, with the following exceptions:

--jobs now takes a keyword ("auto", "HOST_CPUS", "HOST_RAM"), optionally followed by [-|*] <float>. "auto" evaluates to the number of CPUs.

Explicitly setting --jobs > MAX_JOBS logs a warning and sets --jobs=MAX_JOBS. This is what we were previously doing for "auto". It makes sense for these to behave similarly, and simplifies our tests.

Setting --jobs=0 throws an exception. Previously, setting --jobs=0 silently set jobs at 1. This logic was confusing, especially because passing 0 to other flags disables them, and this one behaved differently. Integration tests that used to pass 0 to jobs, which implicitly set jobs at 1, now set jobs at 1 explicitly.

All validity checks and adjustment of the number of jobs now live in the jobs converter. Extra validity checks of the number of jobs are now removed. Now that all of the logic is in the jobsconverter (+ corresponding open source tests), we can eliminate our corresponding internal tests, and integration tests that just rely on logs.

PiperOrigin-RevId: 226507306