Add a check for android_binary depending on multiple android_library rules with databinding in the java same package.
Android databinding generates a class with a fixed name into the Java package of an android_library. This means that two or more android_library targets with the same Java package in the transitive dependencies of and android_binary will lead to class conflicts and mysterious one-version violations. This check checks for this case and gives a specific error message.

RELNOTES: None.
PiperOrigin-RevId: 226515044