For Java binaries add directories containing native libraries to LD_LIBRARY_PATH.
When a Java binary depends on a native shared library, Bazel adds the directory
containing the library to the binary's java.library.path system property.  This
allows the Java runtime to find the library when the binary dynamically loads
it using System.loadLibrary().

However, the library being loaded may depend on another shared library.  When
this is the case, the Unix dynamic linker, ldd, is responsible for loading that
dependency and ldd does not consult java.library.path.  Instead, ldd uses
several techniques to find the dependency.  One of the fallback techniques is
to search for the dependency in the directories listed in the LD_LIBRARY_PATH
environment variable.  This change adds all the directories in
java.library.path to LD_LIBRARY_PATH, so ldd can find the dependency.