Fix a race in ByteStreamUploader.
Register the callback to remove a pending upload from the uploadsInProgressMap after adding it to the map. Before, the callback could execute in the caller's thread as soon as it was registered and vacously remove the pending upload from the map before it was added.

I was able to reproduce this race and verify it has been fixed with:
```
$ bazel test //src/test/java/com/google/devtools/build/lib:remote-tests --runs_per_test 50 --test_filter ByteStreamUploaderTest --test_timeout 15
```