Windows, Bazel client: registry read failure is OK
The Bazel client, when it looks for MSYS2's path
in the registry, will no longer display errors
when it fails to read registry keys or the subkeys
are not what the Bazel client expects.

The added code comments explain the rationale in
detail, but the summary is, the registry may have
been modified and/or MSYS2 may have been deleted
or incorrectly uninstalled, leaving behind stale
registry keys, and Bazel should not stumble over
this circumstance.

Fixes https://github.com/bazelbuild/bazel/issues/6685