Attribute parameterize all starlark attribute transition implementation functions.
The impl functions of these transitions currently take one argument but the final API calls for a second. For now, support both 1-param and 2-param transition impl functions to not break current users.

Though this param looks like the attr param of the starlark rule `ctx` object, it's not exactly the same. For dependency-typed attributes, `ctx` can actually access the providers of those dependencies. For this attr object, we just see them as the representative labels. Current usage of transitions has no need to actually see providers so leaving as a TODO for now and we can reconsider when we find a use case.

PiperOrigin-RevId: 227543951