Simplify config options / fragment registration API.
See ConfiguredRuleClassProvider.java for main change.

Since fragments already have to declare which options they require,
it's redundant to add both the options and the fragments that
use them.

Even worse, addConfig doesn't register *all* options a fragment
requires. It only requires that *any* of the options required
by the fragment are specified. This is imprecise and doesn't
promote a clear API.

PiperOrigin-RevId: 227564551