Go through more-or-less-standard error checking when getting configured targets in tests.
This still doesn't emit all possible errors, but it's a lot closer than we were before.

We could try throwing errors instead, but that makes tons of tests fail. Emitting the errors is good enough, I think. Emitting errors (and not throwing) is what a --keep_going build would do, in any case.

PiperOrigin-RevId: 227692048