Remove BazelServerLogModule and --norotating_server_log startup option.
The --norotating_server_log startup option was intended only as a temporary
"escape hatch"; if it had so happened that the new server log handler caused
problems, we could have suggested to users to use that startup option as a
workaround while they were waiting for us to make a patch release fixing the
issue.

No such problem materialized over the course of the Bazel 0.20 release cycle,
and search results show no evidence of anyone using the startup option. Now that
Bazel 0.21 is out, the temporary "escape hatch" can be removed.

RELNOTES: None.
PiperOrigin-RevId: 227706064