Ensure copts attributes and per_file_copt are handled by LTO backends
A recent bug exposed the fact that copts attributes from BUILD rules are not being propagated to the LTO backend compile invocations for ThinLTO builds. Fix that here by saving this information when creating the compile outputs (which was already saving information for the later LTO steps, such as the correspondence between the full bitcode and the minimized bitcode created by the compile step). Also make this save and therefore propagate to the LTO backends the --per_file_copt applying to that file. Added tests for all these cases (and --copt which were already being applied to the LTO backend but not tested).

Note that nocopts attributes are also not handled properly by LTO backends, however that will be fixed by hlopko@ as part of b/122303926.

RELNOTES: None.
PiperOrigin-RevId: 227859083