Do not use "tail" in bazel_test_test to read a file.
The version of "tail" in Ubuntu 14.04 is too old and prints a weird warning message when it can't recognize the filesystem its input file is saved on (cf. https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=810669). This is the case when using Docker, because "tail" in that version doesn't know about overlayfs yet.

Due to the (harmless) warning message, the assert in the test fails. Just not using "tail" and replacing the assert with expect_log seems fine here and lets the test pass.

RELNOTES: None.
PiperOrigin-RevId: 228149388