Allow pkg_tar's strip_prefix to work on directory output target names.
For build targets that yield a directory output, it is sometimes of
interest to turn them into tarballs, Docker images, etc., that contain
the directory output directly, without having any name prefix to it.

This currently breaks, because setting strip_prefix to the same name as
the directory itself causes build_tar.py to use an empty filename. There
is a check in the code to remove leading slashes, which breaks on empty
strings.

Closes #6980.

PiperOrigin-RevId: 228158067