Have Preconditions check more verbose
I experienced the precondition actually fail when some of the
packages used by the targets were not committed to the repository.
Unfortunately, I can not reproduce it on the simple example, but
I definitely think we should keep the error message in that assertion,
otherwise it;s hard to investigate.