Unify the handling of late-bound and implicit attributes in DependencyResolver.
I also tried merging them with explicit attributes, but it ended up subtly changing error reporting (and thus breaking testLoadingOfImplicitTestSuiteTests()) and something about aspects (and thus breaking extraActionsFromDifferentAspectsDontConflict()).

So in the name of making the perfect not be the enemy of the good, I'll be content with not going all the way. The fact that visibility is special-cased is more painful anyway.

RELNOTES: None.
PiperOrigin-RevId: 228293616