Allow starlark-defined transitions to write build settings.
Previously we just operated on native options (for a few whitelisted users). While this CL allows writing build options, it does not validate that the values the transitions set for build settings are of the proper type. Nor does it allow reading of all build options. That will be coming in a follow up CL. Test added to note these gaps.

RELNOTES: None.
PiperOrigin-RevId: 228513456