Use minimal JDK, which is cached on bazel-mirror.
The minimal JDK is build as part of a buildkite pipeline and I'll add a test that it is identical in a follow-up commit. We cannot just build the minimal JDK during the build since the jlink tool produces is highly non-deterministic.

To build the minimal JDK yourself:
bazel build //src:embedded_jdk_minimal
(This is also what we do in the buildkite pipeline.)

Tested with the downstream pipeline here:
https://buildkite.com/bazel/bazel-at-head-plus-downstream/builds/737#b54cc161-a428-4284-922e-12d4d25d3d1b

This will save ~22MB of the final bazel binary and ~120MB in the
install_base.

Closes #7065.

RELNOTES[INC]: //src:bazel uses the minimal embedded JDK, if you want to
avoid the extra steps of building bazel with the minimized JDK, use //src:bazel-dev
instead.

PiperOrigin-RevId: 228553556