Fix crash bug with the 'visible' query function and non-package_group targets.
The helper functions used by BlazeTargetAccessor were assuming the 'includes' attribute value was well-formed, but in fact it's only during the analysis phase that Blaze errors out if any of the labels in there don't correspond to package_group targets.

Instead, we now ignore such illegal targets when computing visibility in the 'visible' function.

RELNOTES: None
PiperOrigin-RevId: 228557171