Work around breakage in existing_rules() use case
Background: The py_binary/py_test attribute "python_version" is guarded by an experimental flag, --experimental_better_python_version_mixing. It is an error to attempt to set this attribute without also passing in the experimental flag.

When users use native.existing_rules() in macro to programmatically copy a target, they end up setting all the attributes on the new target explicitly, even if they were only implicitly set on the old target. This runs afoul of the above check.

This CL exempts python_version from being visible to existing_rules(). This is a total hack, but it's permissible because python_version is still experimental. We need to remove this hack at or before the point when we flip the experimental flag.

RELNOTES: None
PiperOrigin-RevId: 228641067