Windows: fix new_local_repository correctness bug
The bugfix consists of two parts:

1.  NewLocalRepositoryFunction declares Skyframe
    dependencies on the files in the root of the
    repository. This yields a bunch of FileValues.

2.  RepositoryDirectoryValue stores the FileValues
    and considers them in equals() and hashCode().

Explanation:

new_local_repository creates symlinks to the files
in the root of the repository. When the symlink
targets change Bazel notices it and rebuilds
rules depending on the changed files.

However on Windows the supposed symlinks are
actually copies (because on Windows Bazel cannot
create file symlinks) and Bazel won't notice when
the "symlink" targets change.

The first part of the bugfix ensures that the
SkyFunction associated with the repository rule
notices the file changes, is re-evaluated, and
re-creates the symlinks (i.e. re-copies the files)
as a side-effect.

The second part ensures that the
SkyFunction-computed value will be different
(because some of the FileValues are different) so
change pruning in Skyframe won't believe that the
the repository's SkyValue stayed the same, and
will rebuild downstream SkyFunctions (i.e. rules).

Fixes https://github.com/bazelbuild/bazel/issues/7063

Closes #7076.

PiperOrigin-RevId: 228863875