Refactor FDO artifacts model.
We currently support two kinds of Feedback-Driven Optimization profiles, offered as compiler inputs: one is based on branching information - FDO/AFDO/XAFDO; the other is the cache prefetching profile. The two types are independent from each other.

This change models the FdoProvider as a container for the independent profile artifacts provided as inputs to the compiler.

The refactoring preferred grouping together the various fdo artifacts: their plumbing is identical.

The option to build an instrumented branch FDO binary is modeled separately from this container, as it is not an input artifact.

The scope of this change was just the structural refactoring. A subsequent change would update naming to better reflect the separation (e.g. FdoMode could be renamed more precisely)

PiperOrigin-RevId: 228897528