Leave sandboxfs mounted when --sandbox_debug is on.
This is very useful to manually rerun failing actions caused by a faulty
sandboxfs, and to actually be able to investigate the work directories as
the description of the flag claims.

While doing this, fix the description of the flag to match reality (the
flag causes the work directories to _not_ be deleted, and this has been
the case for a while), and make all sandboxfs-related debugging messages
depend on this flag.

RELNOTES: None.
PiperOrigin-RevId: 229158578