Fix bug in test.xml generation
test-setup.sh was setting a trap to call write_xml_output_file, but
there's no test.log.xml if split xml generation is enabled.

Add a test case for this and comment out a part of the test case that is
not actually working (the test passes because it ignores the exit code
from perl).

Progress on #4608.

PiperOrigin-RevId: 229167973