Remove mount_path_toolchain and associated test.
The archive changed on the GCS bucket without notice, which caused our pre- and postsubmit pipeline to fail. I wanted to get rid of the test for a long time already, as it downloads a massive ~107 MB file just to test that the linux-sandbox can actually mount custom paths (which is a feature I'm not sure anyone actually uses and which I'd like to remove, too).

RELNOTES: None.
PiperOrigin-RevId: 229174952