C++: Removes 'alwayslink' check in LibraryToLinkWrapper
LibraryToLinkWrapper had a precondition check so that alwayslink would not be
true unless a static library or a pic static library was passed. This breaks
rules_go. Removed the check for now, made it a rule warning instead and will
make it a rule error in the future. Then we will be able to add back the
precondition check.

See: https://github.com/bazelbuild/rules_go/issues/1909

RELNOTES:none
PiperOrigin-RevId: 229203335