Automated rollback of commit c54c98ca75690bdddf8ec1725943a8383b593252.
*** Reason for rollback ***

Breaks internal tests

*** Original change description ***

Add the ability to sort manifests in dependency order instead of path order.

Currently, mergee manifests are listed in path order. However, this may result
in erroneous conflicts. Dependency order, with the manifests of dependencies
listed after the manifests of the targets which depend on them, allows for
proper overriding.

Note that this just uses naive link order. Technically, this is incorrect...
But changing these nested sets to use proper link order is a change for another
day. And someo...

***

RELNOTES: None.
PiperOrigin-RevId: 229226334