Fix check-then-act race with rewinding in action execution
An action could probe buildActionMap to see if a shared action ran for
its output, note that one did, and set up a dummy ActionRunner with a
null context. If another action evicts the buildActionMap entry because
of rewinding, and that eviction wins the race with the dummy
ActionRunner, then the dummy ActionRunner will be used, causing a crash.

RELNOTES: None.
PiperOrigin-RevId: 229269872