Automated rollback of commit 0188e88174334173bfd380965fd535a5435e95d2.
*** Reason for rollback ***

Breaks internal project

*** Original change description ***

Verify that an action is either shareable or discovers inputs, but not both.

We don't make the guarantee that the execution phase sees only one instance of a shared action and it's therefore possible to get in an inconsistent state when said action is stateful. The simplest way to avoid this is to disallow stateful shared actions.

There were efforts to make this work but they did not work completely and were complicated so they were abandoned.

This is the second attempt for https://github.com/bazelbuild/bazel/commit/db4871732a180c31e7a8b4eba0e0485a04a3a857 after...

***

ROLLBACK_OF=225525717

RELNOTES: None.
PiperOrigin-RevId: 229388865