Fix SkylarkCallable signature mismatch error messaging
Instead of using the passed *arguments* for the function signature, use the function *parameters*. The parameter display matches that of parameters for user-defined functions.

This is a definite improvement to error messaging; further improvements may be made based on user input (a design discussion document has been circulated to bazel-dev)

Fixes #6160.

RELNOTES: None.
PiperOrigin-RevId: 229430747