Move the Java tools deploy jars refrences to checked-in deploy jars
instead of embedding the deploy jars built at the same time with the bazel binary.

This comes in preparation for moving the Java tools to remote repositories (see #6316).

Moving the tools happens in two phases:
1. Using checked-in deploy jars/binaries instead of those built when bazel was built
2. Moving the checked-in jars/binaries to a remote repository and make bazel reference them from there. This step can only happen under an incompatible flag.

This PR addresses the first point. Doing the move incrementally this way is safer and allows finding bugs earlier. It also allows moving the Java tools code base out of the bazel repository.

We are making this change gradually for each java tool defined in JAVA_TOOLS (found in src/BUILD).

Closes #7114.

PiperOrigin-RevId: 229524951