Make bash processing stricter and refactor a bit
Perl outputs an error if it encounters an invalid utf-8 character in the
input, which means that the current script doesn't actually work, at
least as of perl v5.24.1.

I don't think we currently do anything with the exit code in Bazel, but
at least we can get it and output a warning or even an error, after this
change.

PiperOrigin-RevId: 229562020