Move the label -> configured target transformation to a single place in DependencyResolver.
We do this instead of doing two passes over most of the code with resolving labels to targets first, then targets to configured targets. The goal of this refactoring is twofold:

1. Make the code slightly more comprehensible
2. Have a single location in the code where we can experiment with various smart ways to avoid the bottleneck whereupon the packages of *all* direct dependencies need to be loaded first, and only then can analysis of any direct dependency (including loading its transitive closure) can begin.

The next step is that the single (dependency kind -> label) -> (attribute -> (label, aspect, configuration)) mapping will be split into two:

1. (dependency kind -> label) -> (dependency kind -> (label, aspect, configuration))
2. (dependency kind -> (label, aspect, configuration)) -> (attribute -> (label, aspect, configuration))

where all the information about the configuration transition derived from the originating rule will be reflected in the result of (1) and the remaining part (the rule configuration transition) will be reflected in the result of (2).

(it's not as simple as it looks because we need to do some error handling based on the type of the target we depend on, but it *almost* is)

The only collateral damage is slightly worse error reporting when a dependency through a visibility= attribute is not a package group. There isn't a fundamental reason for that, it's just that I decided that the additional code branches are not worth it.

RELNOTES: None.
PiperOrigin-RevId: 229785096