Map symlink inputs into the sandbox "as is".
If an action decides to specify an input as a symlink, we must respect that
and assume that the action will also map the target of such symlink as an
input.  If it doesn't, the action is not hermetic and breaks when run under
sandboxfs, which is intentional.

This fixes problems building some iOS projects that have trees of symlinks
to headers in their source trees when sandboxfs is in use.  Without this
fix, these failing actions were seeing both the target of the symlink and
the presumed-symlink as actual files, the compiler didn't understand they
were the same thing, and then complained about duplicate symbols.

RELNOTES: None.
PiperOrigin-RevId: 229792273