Add //src:bazel-bin_dev_jdk target.
This builds the minimal JDK locally and uses it as embedded JDK.
We typically use the cached one on mirror.bazel.build, because the jlink
is heavily non-determistic. However, this will make local testing
easier, e.g. when we do upgrades from one JDK to another.

Progress on #6592 and #6314.

RELNOTES: None

Closes #7175.

PiperOrigin-RevId: 229912262