Support --test_filter for our own shell unit tests
Currently, the bash test framework uses command-line args to determine which tests to run. After this change, it observes --test_filter, which Bazel passes to the test as an environment variable (TESTBRIDGE_TEST_ONLY).

Tested manually with Bazel:
$ bazel test //src/test/shell/bazel:bazel_test_test --test_output=all --test_filter=test_runs_per_test_detects_flakes

By inspecting the test logs, we can see that only the one test is run.

PiperOrigin-RevId: 230329936