Always use output redirection in LocalSpawnRunner
Also refactor the LocalSpawnRunner to use the lower-level
SubprocessBuilder API rather than the Command API. This is in
preparation for migrating to ListenableFuture in order to implement the
new async SpawnRunner API. Also, the SubprocessBuilder is a much better
API than Command.

Note that this means that stdout / stderr cannot be written to Path
objects that do not correspond to an actual file system.

PiperOrigin-RevId: 230336761