C++: Deletes CcLinkingInfo
This is the last CL in the CcLinkParams nightmare. Instead of doing the last 3 or 4 incrementally. We have decided to do one big CL where it gets deleted. These CLs where there is back and forth conversion from the old and the new API are error prone and have been rolled back quite frequently. The errors are introduced in the conversion code. By completely removing CcLinkingInfo and CcLinkParams, these errors go away and the refactoring is finished.

All tests pass and there is no memory regression.
Tested with downstream projects:
https://buildkite.com/bazel/bazel-at-head-plus-downstream/builds/767

RELNOTES:none
PiperOrigin-RevId: 230347403