Revert "Enabling Bazel to generate input symlinks as defined by RE API V2."
This reverts commit baa1786fe6ac01511bb15689a81b2978f39686da.

The symlink resolution in this change is broken, as it does not take into
account parent symlinks. Consider the following structure on the filesystem:

a/d/file
a/b/c/symlink -> ../../d/file

And action inputs as follows:

a/d/file
(f -> a/b/c)/symlink -> ../../d/file

with (f -> a/b/c) denoting that f is a symlink to directory c.

This change then builds the following merkle tree:

a
  d
    file
f
  symlink -> ../../d/file

My guesstimate is that there are a number of additional problems with
this change and we should think hard about them when attempting to
roll foward this change in the future. See for example Bazel's symlink
resolution in FileFunction [1].

A real world example of this error is: #7212
Fixes #7212

[1] https://source.bazel.build/bazel/+/master:src/main/java/com/google/devtools/build/lib/skyframe/FileFunction.java;l=114?q=FileFunction