Avoid unnecessary collection copy in NestedSetBuiler.wrap
NestedSetBuiler.wrap is doing a pre-emptive copy of the Iterable that is wasteful. This could be avoided for plenty of cases and that's what the CL does.

Since this is a central collection, it is worthwhile to check performance implications of this but I don't know if it is easy to do so. Being said that I am not worried about any potential negative impact.

PiperOrigin-RevId: 230390746