remote: don't upload failed action outputs. Fixes #7232
For historical reasons we are currently uploading the outputs
of failed actionsi (exit code != 0). This was mostly so that
test.log and test.xml of failed tests would be uploaded when
BES was enabled. However, we have had local file uploads in
BES for a while now and thus this (unexpected) feature is now
obsolete.