remote: quit early if no remote options specified
If neither remote caching nor remote execution is enabled then
quit the remote initialization logic early. The current null
check on the RemoteOptions is not sufficient as it only works
if a none build command is used.

Also, don't create a remote logs directory and don't print
an invocation id if remoting is disabled.

Closes #7198.

PiperOrigin-RevId: 230573618