Update bootstrap documentation to include `--host_javabase`
Our instruction on how to install bazel from source tell to install a
JDK on the host machine. Then the standard expectation is that this
is used, which is also the standard workflow for packaging software.
To comply with this expectation, the `--host_javabase` argument
has to be set. Fix our documentation accordingly.

Discussed on #6656.

Change-Id: Ib64b35d0b82fee83fcd9282519e16807588ed8a7
PiperOrigin-RevId: 230694090