Vendor jaxb library in our third_party.
We want to upgrade the embedded JDK to version 11 (see #6592). However,
some Java EE modules were removed from 10 to 11:
http://openjdk.java.net/jeps/320

From these modules, we need java.xml.bind for the android_common
dependency which is used by the android_sdk_repository rule.
This will be starlark-ified some time in the future and then will also
move the tools to an external repository making it possible to remove
this and the android_common dependency from Bazel. But since the
migration is not imminent, this commit is the work around to unblock the
upgrade to JDK11.

RELNOTES: None