Respect the max open files set by the Blaze client in the server.
As it turns out, the JVM has a feature to unlimit its own view of what the
maximum number of open file descriptors should be.  This is controlled via
the MaxFDLimit option, which is enabled by default.

The problem is that, on macOS, MaxFDLimit restricts its upping of the open
files count to OPEN_MAX (per the setrlimit manpage), which is 10k.  This
limit is much smaller than what the system truly allows per process.

In unknown commit, I added logic to the Bazel client to unlimit its own
resources at startup time to the real limits of the system.  I obviously
expected those new limits to propagate to the server... but that's not the
case because of the above.

Fix this by disabling the MaxFDLimit option in the server's JVM and add a
test to ensure that the high limit computed by the client propagates to
the actions run by Bazel.

RELNOTES: None.
PiperOrigin-RevId: 230716686