Clean up sandbox scratch directories as we go.
When using sandboxfs, we were leaking the per-spawn scratch directory
created outside of the sandboxfs mount point.  This was not a problem
per se because the SandboxModule would clean up the whole subtree during
build completion... but on large builds, this final cleanup could take
many, many, many minutes because the scratch tree ends up being gigantic.

Fix this by cleaning up each spawn's scratch directory during the
cleanup of that spawn.  This way the cost is spread throughout the build
and we don't bottleneck on disk accesses at the very end.  This is the
same as the symlink-based sandboxing runner does.

Note that creating these complex tress outside of the sandboxfs instance
defeats the whole point of sandboxfs... I should rethink how to avoid
this, but not as part of this change.

RELNOTES: None.
PiperOrigin-RevId: 230732316