Fix a race in ByteStreamUploader deduplication.
The callback to add a completed digest to the uploadedBlobs set was added to the result future of uploadBlobAsync. The non-deterministic ordering of future callbacks meant a client of ByteStreamUploader could observe a completed upload before the digest was added to uploadedBlobs. The solution is to chain futures so that the digest is placed in uploadedBlobs before the final future fires.

I noticed this race caused a test failure in CI: https://storage.googleapis.com/bazel-buildkite-artifacts/f1b54300-cf87-4516-8767-410b93ab1ee9/src/test/java/com/google/devtools/build/lib/remote-tests/test.log

It also was readily reproducible for me with this command:
$ bazel test //src/test/java/com/google/devtools/build/lib:remote-tests --runs_per_test 50 --test_filter deduplicationOfUploadsShouldWork
```
There was 1 failure:
1) deduplicationOfUploadsShouldWork(com.google.devtools.build.lib.remote.ByteStreamUploaderTest)
expected: 1
but was : 2
        at com.google.devtools.build.lib.remote.ByteStreamUploaderTest.deduplicationOfUploadsShouldWork(ByteStreamUploaderTest.java:731)
        at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
        at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:62)
        at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
```

We can also simplify things by removing the second parameter of startAsyncUpload. This in turn permits us to delete a Retrier.executeAsync overload.